resume_qna_bot/requirements.txt
langchain
openai
pdfplumber
streamlit
python-dotenv
